refactor: make one instance of copy, clear buttons
